Create Dataset_and_trained_models.txt
# Drowsiness-Detection-System-Using-NeuralNetworks-BasicCNN-VGG16
The objective of this intermediate Python project is to build a drowsiness detection system that will detect that a person’s eyes are closed for a few seconds. This system will alert the driver when drowsiness is detected.  In this Python project, we will be using OpenCV for gathering the images from webcam and feed them into a Deep Learning model which will classify whether the person’s eyes are ‘Open’ or ‘Closed’. The approach we will be using for this Python project is as follows : Step 1 – Take image as input from a camera.  Step 2 – Detect the face in the image and create a Region of Interest (ROI).  Step 3 – Detect the eyes from ROI and feed it to the classifier.  Step 4 – Classifier will categorize whether eyes are open or closed.  Step 5 – Calculate score to check whether the person is drowsy.
FoR DataSet and haar models and trained models, check link https://drive.google.com/drive/folders/1xnyhgd9kSf7uMNv3GAPkPkRfjy0mBKub?usp=sharing